requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
multidict==6.0.5
SQLAlchemy==2.0.27
typing_extensions==4.9.0
yarl==1.9.4
